refactor: remove unused JWT middleware, add video API for uploads, and enhance cloudinary upload functionality

+++ server/src/utils/cloudinary.js
@@ -8,41 +8,43 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadOnCloudinary = async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath, resourceType = "auto") => {
   try {
     if (!localFilePath) return null;
-    const response = await cloudinary.uploader.upload(localFilePath, {
-      resource_type: "auto",
+    const fileStream = fs.createReadStream(localFilePath);
+
+    return new Promise((resolve, reject) => {
+      const uploadStream = cloudinary.uploader.upload_stream(
+        { resource_type: resourceType, chunk_size: 100 * 1024 * 1024, timeout: 60000 },
+        (error, result) => {
+          if (error) {
+            reject(new Error('Cloudinary upload failed'));
+          } else {
+            if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath); // Delete local file after upload
+            resolve(result);
+          }
+        }
+      );
+
+      fileStream.pipe(uploadStream);
     });
-    fs.unlinkSync(localFilePath);
-    return response;
   } catch (error) {
-    // remove local file from temp
-    fs.unlinkSync(localFilePath);
-    console.log(error);
+    if (fs.existsSync(localFilePath)) fs.unlinkSync(localFilePath);
     return null;
   }
 };
 
 const extractPublicId = (url) => {
   const parts = url.split("/");
-  const fileName = parts.pop(); // Get the file name with extension
-  const publicId = fileName.split(".")[0]; // Remove the file extension
-  return publicId; // Return only the public ID
+  const fileName = parts.pop();
+  return fileName.split(".")[0]; // Return only the public ID
 };
 
-const deleteOnCloudinary = async (filePath) => {
+const deleteOnCloudinary = async (filePath, resourceType = "auto") => {
   try {
     if (!filePath) return null;
-
-    // Extract public ID
     const publicId = extractPublicId(filePath);
-
-    // Delete from Cloudinary
-    const response = await cloudinary.uploader.destroy(publicId, {
-      resource_type: "image",
-    });
-
+    const response = await cloudinary.uploader.destroy(publicId, { resource_type: resourceType });
     return response;
   } catch (error) {
     return null;
